Added manpage flag to Makefile and edited INSTALL to mention how to configure Makefile manually.

+++ Makefile
@@ -10,6 +10,10 @@
 
 BIN_LOCATION=/usr/local/bin
 
+# Change the value of CREATE_MANPAGE to zero (0) if you don't want to install a manpage.
+
+CREATE_MANPAGE=1
+
 # A list of all the files in the application. These are the files that 'make' moves about
 
 scripts = fetchheaders.py miscClasses.py imapServer.py
@@ -24,14 +28,14 @@ manFiles = fetchheaders.1.gz
 
 install:
 
-	@echo "\nInstalling the application files. Run 'make install_binary' to install the binary and manpage. Run 'make help' for a list of options.\n"
+	@echo -e "\nInstalling the application files. Run 'make install_binary' to install the binary and manpage. Run 'make help' for a list of options.\n"
 	
 # Check if ~/.fetchheaders folder exists. If not create it :
 
 	@if [ ! -d ~/.fetchheaders ]; then \
 	\
 		mkdir ~/.fetchheaders ; \
-		echo "mkdir ~/.fetchheaders\n" ; \
+		echo -e "mkdir ~/.fetchheaders\n" ; \
 	fi
 
 # Copy .py scripts to ~/.fetchheaders :
@@ -44,7 +48,7 @@ install:
 
 # Check if configuration and specificiation files exist. If not copy them.
 
-	@echo "\nChecking if configuration files exist and copying them if they do NOT.\n"
+	@echo -e "\nChecking if configuration files exist and copying them if they do NOT.\n"
 
 	@for file in $(configs); do \
 	\
@@ -61,19 +65,22 @@ install:
 
 install_binary:
 
-	@echo "\nInstalling the binary to $(BIN_LOCATION) and the manpage to /usr/share/man/man1/. Should be run as root.\n"
+	@echo -e "\nInstalling the binary to $(BIN_LOCATION) and the manpage to /usr/share/man/man1/. Should be run as root.\n"
 
-# Copy manpage. This process will require root access since it accesses the /usr/... folder tree:
+# Copy manpage if the CREATE_MANPAGE flag is equal to 1. This process will require root access since it accesses the /usr/... folder tree:
+	
+ifeq ($(CREATE_MANPAGE),1)
 
 	@cp fetchheaders.1.gz /usr/share/man/man1/
 	
-	@echo "\ncp fetchheaders.1.gz /usr/share/man/man1/"
+	@echo -e "\ncp fetchheaders.1.gz /usr/share/man/man1/"
+endif
 
 # Create symbolic link to the fetchheaders main script
 
 	@ln -s ~/.fetchheaders/fetchheaders.py $(BIN_LOCATION)/fetchheaders
 
-	@echo "\nln -s ~/.fetchheaders/fetchheaders.py $(BIN_LOCATION)/fetchheaders"
+	@echo -e "\nln -s ~/.fetchheaders/fetchheaders.py $(BIN_LOCATION)/fetchheaders"
 
 
 
@@ -82,7 +89,7 @@ config:
 
 # Installs the configuration and specification file, possibly over-writing ones that already exist.
 	
-	@echo "\nInstalling (and possibly over-writing) configuration files.\n"
+	@echo -e "\nInstalling (and possibly over-writing) configuration files.\n"
 
 	@for file in $(configs); do \
 	\
@@ -101,7 +108,7 @@ purge: uninstall
 
 # Remove the configuration files and then run the "uninstall" recipe
 
-	@echo "\nRemoving configuration files.\n"
+	@echo -e "\nRemoving configuration files.\n"
 
 	@-for file in $(configs); do \
 	\
@@ -115,7 +122,7 @@ purge: uninstall
 
 uninstall:
 
-	@echo "\nRemoving the fetchheaders scripts and manpage but NOT the configuration files (run 'make purge' to remove configuration files as well).\n"
+	@echo -e "\nRemoving the fetchheaders scripts and manpage but NOT the configuration files (run 'make purge' to remove configuration files as well).\n"
 
 # Removing 'fetchheaders' scripts but not configuration files
 
@@ -135,23 +142,28 @@ uninstall:
 
 # Removing 'fetchheaders' symoblic link, requires root access.
 
-	@echo "rm $(BIN_LOCATION)/fetchheaders\n"
+	@echo -e "rm $(BIN_LOCATION)/fetchheaders\n"
 
 	-rm $(BIN_LOCATION)/fetchheaders
 
-# Remove fetchheaders manpage
 
-	@echo "rm /usr/share/man/man1/fetchheaders.1.gz\n"
+# Remove fetchheaders manpage if the configuration flag indicates that it has been installed in the first place.
+
+ifeq ($(CREATE_MANPAGE),1)
+
+	@echo -e "rm /usr/share/man/man1/fetchheaders.1.gz\n"
 
 	-rm /usr/share/man/man1/fetchheaders.1.gz
+	
+endif
 
 
 
 help:
 
-	@echo "\nOptions:" 
-	@echo "\n\tinstall - Installs the applications scripts and the configuration files unless the configuration files already exist in which case it DOES not over-write them."
-	@echo "\n\tinstall_binary Installs the fetchheaders binary to BIN_LOCATION and the manpage to /usr/share/man/man1/. Should be run as root."
-	@echo "\n\tuninstall - Remove application scripts, binary and manpage. Leave the configuration files. Run as root if you wish to remove the binary and manpage."
-	@echo "\n\tpurge - Remove ALL application files and folders including the configuration files. Run as root if you wish to remove the binary and mangpage."
-	@echo "\n\tconfig - Install the configuration and specification files, over-writing existing ones. Use to create clean configuration and specification files."
+	@echo -e "\nOptions:" 
+	@echo -e "\n\tinstall - Installs the applications scripts and the configuration files unless the configuration files already exist in which case it DOES not over-write them."
+	@echo -e "\n\tinstall_binary Installs the fetchheaders binary to BIN_LOCATION and the manpage to /usr/share/man/man1/. Should be run as root."
+	@echo -e "\n\tuninstall - Remove application scripts, binary and manpage. Leave the configuration files. Run as root if you wish to remove the binary and manpage."
+	@echo -e "\n\tpurge - Remove ALL application files and folders including the configuration files. Run as root if you wish to remove the binary and mangpage."
+	@echo -e "\n\tconfig - Install the configuration and specification files, over-writing existing ones. Use to create clean configuration and specification files."
